[Move] `ToolCallParsedResult` code location
**Major**:
- From in `mcp_tool_execution.py` to `data_structures.py`.
- Adapted the import statements codebase-wise.
